requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
Flask-Login
eventlet
Pillow
gunicorn
python-socketio
python-engineio
python-dotenv
giphy-client